Code Review Fixes

+++ src/common/midi_event.py
@@ -5,4 +5,4 @@ import mido
 @dataclass(order=True)
 class MidiEvent:
     event: mido.Message=field(compare=False) # Don't use this when sorting in a priority queue
-    timestamp: int # TODO: Better time class?
+    timestamp: float # Time.time() returns seconds as float
